Add base KeyChainParser for passthrough modes
